Updated Wrapper
The call to the API now ignores all columns except for the address (before all the columns were concatenated), and returns an .csv file with just the numerical coordinates added to useless info
